Refine user authentication for D7 and Facebook
This is the first part of migration to the new API of D7 site. The
logon and logout flow now conform to the new D7 site, as users can
now sign in with both Facebook and Drupal accounts of TaiRON.

As for the authentication, Drupal7 handles CSRF well and the app
has already been dealing with the token. Furthermore, the services
fboauth module on D7 makes it really simple for the app to sign in
with Facebook. Authentication is exactly checked before each login.

The new logon page is designed with some reference as following
CSS: http://codepen.io/aklatzke/pen/JwcCx
Icon: http://andymatthews.net/code/jQuery-Mobile-Icon-Pack/builder/